Move tablet management code from StorageEngine to TabletManager
