+++ be/src/olap/schema_change.cpp
@@ -1687,8 +1687,7 @@ OLAPStatus SchemaChangeHandler::_create_new_tablet(
         // 4. Register tablet into store, so that we can manage tablet from
         // the perspective of root path.
         // Example: unregister all tables when a bad disk found.
-        res = StorageEngine::get_instance()->register_tablet_into_root_path(
-                new_tablet.get());
+        res = new_tablet->register_tablet_into_dir();
         if (res != OLAP_SUCCESS) {
             OLAP_LOG_WARNING("fail to register tablet into root path. "
                              "[root_path='%s' tablet='%s']",
