+++ be/src/olap/tablet.cpp
@@ -298,6 +298,30 @@ EXIT:
     return res;
 }
 
+bool Tablet::can_do_compaction() {
+    // 如果table正在做schema change，则通过选路判断数据是否转换完成
+    // 如果选路成功，则转换完成，可以进行BE
+    // 如果选路失败，则转换未完成，不能进行BE
+    this->obtain_header_rdlock();
+    const PDelta* lastest_version = this->lastest_version();
+    if (lastest_version == NULL) {
+        this->release_header_lock();
+        return false;
+    }
+
+    if (this->is_schema_changing()) {
+        Version test_version = Version(0, lastest_version->end_version());
+        vector<Version> path_versions;
+        if (OLAP_SUCCESS != this->select_versions_to_span(test_version, &path_versions)) {
+            this->release_header_lock();
+            return false;
+        }
+    }
+    this->release_header_lock();
+
+    return true;
+}
+
 OLAPStatus Tablet::load_indices() {
     OLAPStatus res = OLAP_SUCCESS;
     ReadLock rdlock(&_header_lock);
@@ -2180,6 +2204,10 @@ OLAPStatus Tablet::test_version(const Version& version) {
     release_header_lock();
 
     return res;
+}
+
+OLAPStatus Tablet::register_tablet_into_dir() {
+    return _store->register_tablet(this);
 }
 
 }  // namespace doris
